Remove translatePosition from transform interface
